+++ R/run_shiny_app.R
@@ -12,9 +12,11 @@
 #'
 #' @export
 run_shiny_app <- function(port = 8080) {
-    app_path <- system.file("shiny_app", package = "iaoreader")
+    ui_path <- system.file("shiny_app", "ui.R", package = "iaoreader")
+    server_path <- system.file("shiny_app", "server.R", package = "iaoreader")
+    www_path <- system.file("shiny_app", "www", package = "iaoreader")
 
-    if (app_path == "") {
+    if (ui_path == "" || server_path == "" || www_path == "") {
         stop(
             paste(
                 "Internal package files are missing.",
@@ -24,5 +26,19 @@ run_shiny_app <- function(port = 8080) {
         )
     }
 
-    runApp(app_path, port)
+    # Note: sourcing those files allows avoiding namespace issues. If the app
+    #       would be run directly with:
+    #           runApp(system.file("shiny_app", package = "iaoreader"))
+    #       then the functions imported by the package could not be used within
+    #       the application due to the Shiny creating its own namespace.
+    #       By sourcing those files all functions imported by the package are
+    #       available withing the application.
+    source(ui_path, local = TRUE, chdir = TRUE)
+    source(server_path, local = TRUE, chdir = TRUE)
+
+    # Manually attaching the resources to /www static URL.
+    addResourcePath("www", www_path)
+    app <- shinyApp(ui(), server)
+
+    runApp(app, port)
 }